coffin-rewrite/requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiomisc==17.5.26
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
async-tools-ext==0.0.1
asyncer==0.0.8
asyncpg==0.30.0
asyncspotify==0.12.1
asyncstdlib==3.13.0
attrs==24.2.0
av==12.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
bokeh==3.1.0
boltons==24.0.0
braceexpand==0.1.7
Brotli==1.1.0
browser-cookie3==0.19.1
bs4==0.0.2
cashews==7.3.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chrome-fingerprints==1.1
click==8.1.7
cloudpickle==3.1.0
cloudscraper==1.2.71
color_processing==1.1
coloredlogs==15.0.1
colorgram_rs==0.1.0
colorlog==6.8.2
contourpy==1.3.0
cryptography==43.0.3
cssselect==1.2.0
ctranslate2==4.5.0
curl_cffi==0.7.3
cycler==0.12.1
dacite==1.8.1
dask==2024.3.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
discord-ios==1.1.1
distributed==2024.3.1
docopt==0.6.2
durations-nlp==1.0.1
exceptiongroup==1.2.2
executing==2.1.0
fast_string_match==0.1.5
fastapi==0.115.5
faster-whisper==1.0.3
fastjsonschema==2.20.0
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
gpt4free==0.0.0
greenlet==3.1.1
gTTS==2.5.3
h11==0.14.0
h3==4.1.2
html5tagger==1.3.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
humanize==4.11.0
idna==3.10
ImageHash==4.3.1
import_expression==1.1.5
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.4
jishaku==2.5.2
jmespath==1.0.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
kaleido==0.2.1
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
limits==3.14.1
locket==1.0.0
logging-journald==0.6.9
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
munch==4.0.0
mutagen==1.47.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netifaces==0.11.0
num2words==0.5.13
numpy==2.0.2
onnxruntime==1.19.2
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
playwright==1.48.0
plotly==5.24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==4.4.0
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pycurl==7.45.3
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.2.0
quantulum3==0.9.2
RapidFuzz==3.10.1
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
sanic==24.6.0
Sanic-Cors==2.2.0
sanic-routing==23.12.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sgmllib3k==1.0.0
simpervisor==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tabulate==0.9.0
tblib==3.0.0
tenacity==9.0.0
terminado==0.18.1
timezonefinder==6.5.4
tinycss2==1.4.0
tokenizers==0.20.1
toolz==1.0.0
tornado==6.4.1
tqdm==4.66.5
tracerite==1.1.1
traitlets==5.14.3
tuuid==1.0
typeguard==4.4.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
Unidecode==1.3.8
unidecode-rs==0.1.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
w3lib==2.2.1
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
whisper==1.1.10
wordcloud==1.9.3
wrapt==1.17.0
xxhash==3.5.0
xyzservices==2024.9.0
yarg==0.1.9
yarl==1.16.0
yt-dlp==2024.10.22
zict==3.0.0
zipp==3.20.2
